+++ app.js
@@ -20,7 +20,7 @@ app.use(require('./podcast/routes'))
 app.use(require('./rss/routes'))
 app.use(require('./wallet/routes'))
 app.use(require('./rippled/routes'))
-app.use(require('./advertising/routes'))
+app.use(require('./advertising/advertisements/routes'))
 app.use(require('./advertising/adcampaign/routes'))
 
 app.listen(3000, () => {
